Extract MiddlewareBuilderPipelineExtension internal and private methods to a dedicated PipelineProcessor class
